Code style cleanup and small fixes, up to Creature.cpp,h
- Added missing copyright headers
- applied the code convention, mostly respecting 4 spaces as indentation.
- Made sure members were initialized
- Added m in front of members, eg: mSpeed.
- Grouped functions together when they were mixed with members.

I also remove two todo since I did one (make sure the constructor was complete)
and one was not relevant anymore (HP, MP, and EXP are all three using decimal values
for computation, and thus require a double, so no need to port them to an unsigned int.

I also made the creatures start with full health as it will prevent them from dying
instantly when a player put them in front of enemies at game start,
which is quite not what the player is expecting.

The creature class is also using gotos, and I made them unindented, to make them more
proeminent. I'll try to remove the use of them in other commits.