Streamlined a bit the Gui handling and fixed an app crash.
Now, every game mode is taking care of loading its own gui sheet.

The game mode itself also start the server now, as this permits
to see to load the game map before starting it and is more overall
logical.

I also removed the need of a mode manager reference in the gui
handler.