Made the camera start position be set by the application mode.
I removed the need to set up some hardcoded map size
in the GameMap constructor.
Now, each application mode can set the camera starting position.

As a result, the game mode will now start at the player's temple
coordinates, and the editor mode will start in the middle of the map.

We could later even set a start position for each team in the map data.

I also turned the culling manager into a component of the camera manager
and made the camera initialisation be done with the camera manager.