Started cleaning up the game map related classes.
As a first step, I removed a few dead members, fixed the behaviour of clearTiles()
du to missing initialization, and merged back the useful functions to the tile container.

The useful functions in the tile container modificator were belonging to the tile container.
The others were broken and unused, and I couldn't find any use for those atm, so I dropped
the now useless modificator class.

This permitted a few simplifications while on it.