Made it so a server message is only sent to the client when starting or stopping the battle music.
As discussed with @hwoarangmy, I moved the battle music timing logic into the serve game map,
and the battle music timer into the player class, as it is a player thing, as righteously said
by hoarangmy.

I also normalized the battle music to avoid hearing glitches in the sound as reported by Akien.