Made the first treasury tile cost nothing.
This handled that way in the server, and the game mode text overlay shows a price of 0
for the first tile in that case.

I also made it so the text overlay is now colored in red when the player hasn't got enough
gold to build a room or a trap.

Fixes #87