Made the level loading work even on system-wide installation.
Music and sounds are already working, but scripts aren't yet.
I made the server send a filename relative to the installation path to the client(s).

Due to the sharable nature of levels, we might later want to just send a filename + SHA/MD5
and let the client search through the bundled levels + user custom ones.