Added several new stats in the creature config file.
Most of them are now used when leveling up.

Also, the actual creature movement speed depending on the terrain is used
instead of its definition base value.

I also added a path() proxy function when called for creature for simpification
purpose.

For now, only the Physical attack/defense are used when fighting.
What's left to add is handling the natural range of the creature and split
physical/magical attack/defense.

Part of #152

Stats added in file.
    DigRate/Level
    ClaimRate/Level
    GroundSpeed/Level
    WaterSpeed/Level
    LavaSpeed/Level
    PhysicalAttack
    PhysicalAtk/Level
    MagicalAttack
    MagicalAtk/Level
    PhysicalDefense
    PhysicalDef/Level
    MagicalDefense
    MagicalDef/Level