Add support to put equipment in a config file In this patch : - Changed the damage computation so it is not needed to compute for every entity that can attack - In the level file, we can only add the weapon name, not its definition - Weapon is not a GameEntity anymore, just a Creature attribute - In the level file, creature classes and equipments can be tuned (or even created) Fixes #138
To test it, I've added in the skirmish 2v2 test map a knight with a tuned LongSword, an adventurer with a level created staff.
I've also set the Kobold base attack to 4 instead of 1 and there is a level specific creature : BigDragon.
Note that there is a BigDragon near the player dungeon and BigDragon can spawn in the top right player portal.