Separated tiles used for defining common actions from the ones used to see enemies positions.
This is fixing #214 to me.

Now workers are much more efficient while creatures won't know of each others
until they face one another.

There is another bug you can see in Scream in the Dark though.
The AI being on the right of the map have worker that don't do anything.
As far as I've looked, it's because they are trying to reach undiggable tiles on the right.

I'll fix this in another commit.