Added possibility to setup alliances in the lobby fixes #310
The available team ids that a seat can use are defined in the level file
Note that the order matters as the first one in the list will be selected by default.
With this logic, some seats can be sealed while others can be setup in the same level.
Examples to test are used in multiplayer small test map FFA and skirmish 2v2 small test map.