Added a replay menu
Please note that this feature is pretty experimental as it relies on timestamps where the server/client relies on message synchronization
But because the client in replay is the most simple, timings should not be a too big problem
Moreover, the replays takes there information from the level map. That means that if the level file is modified, there can be weird issues (same as wrong synchronization between server and client level)

But IMHO, it is a cool feature and it can help a lot debugging weird issues (like ghost tiles) so it is worth the price ^^

Later, it would be nice to have options to:
- Save only last replay (should be default)
- Do not save any replay
- Save all replay (like it is currently done)
Because on long games, replay could be quite big (for 10 minutes, it was more or less 10Mb from what I've seen)

Also note that every server related events won't be available (like the creature stats info unless it has been displayed at the same time).
Moreover, because everything is timestamped, if you pause the game, you will see it pause during replay ^^