Added spawn conditions to the Hero faction
I didn't make Gnome have spawn conditions on purpose, and I tried to follow the logic in spawn conditions so that:
Dwarf1 and Adventurer creatures come early but with spawn conditions.
Dwarf1, 2 and 3 make having a forge something more important for heroes.
Knights and Wizard dislike Dwarf3 creatures so you can either choose to first build a forge and get plenty of dwarves, or a library and try to get Wizards and (Big)Knights.

Part of #313