Fixed the tooltips offset.
This was coming from the fact of using the OS cursor only, as CEGUI considers
it being 0x0 in size, and then computes the offset badly in that case.
I thus added an empty 12x12 cursor and used it without hiding the system one.

I also removed a obsolete comments and explained another one.

Fixes #337