Preparation commit to later deal with #134
For now, I simply aligned the available console command, their short and long doc.

I also removed the non functional, duplicate or IMHO now useless ones and renamed
a few to be less confusing.

The actual goal is to load the command name, its short and long descriptions
within the same file, but it will be for the next commit.

If you see a command tthat shouldn't have to be removed or should be removed, feel free to tell. :)

Part of #134