Fixed a crash when active objects were removed during AI turn
I don't know if this case was possible before AI tried to repair rooms but now, it may happen if room objects are removed during room absorbtion. The bug is due to the fact that active objects were removed after active object turn (which happens before AI turn). Then, if an active object is removed during AI turn, it will be deleted but not removed from the active queue list causing a crash when its turn comes