Rename tentacles for clarity and fix unused 2nd model
Because of a typo in creatures.cfg, the two tentacles were using the
same mesh. This commit fixes it and also renames the tentacles meshes
and creature definitions for more clarity. "Tentacle" becomes
"TentacleAlbine" and "Tentacle2" becomes "TentacleGreen". The skeleton
used by the two meshes is exactly the same, so the duplicate one is
removed and both use Tentacle.skeleton.