Added more entities animation fallbacks.
And made the animation names set in a common set.

Now, when either the dig or claim animations are missing, the attack one will be used
if possible.
When the sleep animation is missing the die one will be used as well.

In any case, the idle animation will be the default fallback when everything fails.

This is fixing Hero worker claiming/digging animation and made a few more creatures
behave more naturally when sleeping.