Simplified modes handling.
Now only one mode will be loaded at a time, the Fpp and Console modes
are now simple components of the GameEditorModeBase class.

I also removed a lot of function aliases in the mode manager and made other smaller cleanups.

What's left is to rename the MenuMode class to MainMenu and the ConsoleMode class
to GameEditorModeConsole.

Biggest part of #670