Fixed the console auto-completion when using tab.
The console editbox was executing the prompt command
when this event was happening: CEGUI::Editbox::EventTextAccepted
But this CEGUI event is also triggered when tab is pushed, breaking the auto-completion.

This is now fixed.

I also added support to use the keypad enter key as well as the standard enter key
for chat/console.

Fixes #757