Use GPLv3+ shortname instead of GPL 3.0
This helps make it clear that we use the "GPL version 3 (or any later version)"
and not a strict GPLv3. This is not a relicensing as "(or any later version)"
is already present in our code headers. See how hard the relicensing would be
otherwise: https://fr.dolphin-emu.org/blog/2015/05/25/relicensing-dolphin/