Added support to play different sounds for creatures. Also play everysound the same way.
Sounds are defined by their type (the first directory in the Sounds folder) - hardcoded
Then, there is their family. The family represents the directories under the main folder. There can be more than 1.

To play a sounds, we use its type and family. At game load, every sound in the sound folder is read and inserted in a map including its path. When a sound is played, a .ogg file is randomly picked in the corresponding folder and played.

fixes #126