Fixed compilation link error when linking lib intl (from gettext) with MinGW64. AFAIK, it is required by bfd for MinGW32.
Note that from what I've seen, there is no way to differentiate MinGW32 and MinGW64 with CMake variables. Thus, to do so, I'm testing the compiler version. If it is 4.8.1 (latest MinGW32), we assume it is MinGW32. Otherwise, we assume it is MinGW64. That may break compilation for older MinGW32 versions (even if 4.8.1 is used for more that 1 year) but since it is not tested, we are not even sure it ever worked.
In any case, from what I've seen on internet, MinGW64 is prefered to MinGW32 because it is much more up to date (ATM, it uses gcc 5.1.0).