Improve BASH unit test script to remove hardcoded test names
The tests are now named with a $basename-LL-$test convention where LL is the name of the level that should be run in the server, if any (00 if none).
Also fixed final check to know if some tests failed that would not work on strings with spaces (i.e. more than one failed test).