Added support to load user levels.
The different skirmish/multiplayer menu handle new "Custom *" level types
which are loaded in:
<userGameDataPath>/levels/skirmish
<userGameDataPath>/levels/multiplayer
respectively.

I also fixed a bug where the level type list entries would be kept in memory and be listed
as much time as the corresponding window is opened.

fixes #383

What's next is to add support to create maps and save them correctly in the map editor.