Allow to launch a saved game in server mode
ATM, the server only can save the game. That means that a game should be created locally to be allowed to be saved. To be really functional, we should allow to save game while playing with the master server. However, this feature would require many changes and will be for later. For now, we should use it for tests only.