Do not process entities overlays that are not in scene
In any case, when the creatures are not visible, the overlays should be out of the screen and, thus, not visible. But in any case, no need to compute and display the overlays if they are not in scene.

Since we now use Ogre to check if the creatures are visible and hide the overlays if not, I will assume it
fixes #1146

If not, we can open another issue