Ogre: Fix font loading with Ogre 1.11
CEGUI initialization failed as it seems it can't resolve fonts/*.ttf if
we have both gui and gui/fonts in the GUI group... Removing the folder
name fixes it.

Ogre 1.11 is now more strict about resource loading, so if we use different
groups from General, we need to be explicit about them.